Two bugs fixed here.
First, since shops had been moved to no-inventory, I altered the give command to switch to 'sell' for items.  This turns out to be a bad idea  and breaks giving to other players and all sortos of stuff. Instead of trying to amke this work - it was a hack afterall- Ill back out this change.  We can revisit if we get alot of newbs handing objects to shopkeepers and thinking itll end up in the shop list.

Second, the fish store and otehr 'dark' stores cause alot of player issues which they constantly complain about.  So there are two problems: 1- when listing objects are always shown, and when buying they are found using 'visibility' settings.  This means you can list and not buy items.  2- we dont tell you when we tried to let you buy something but you couldn't see it.  So that's now fixed.