Fix error handling in parseFollowerRentFile
(Caught by clang's static analyzer) Looks like error handing started with
'break'ing out of the main for-loop, only this breaks down when you have inner
for-loops.  So, switch to C++ exception handling.

I picked an exception over alternatives because it nicely avoided changing the
indent level of a whole batch of code.