Portal Buff
Intent: Increase the timer on the 'Portal' skill from 5 minutes to 10 and a max number of uses from 10 to 20. 

If I have done so incorrectly, please push back.

Reason: I feel the nerf that took it down to this level originally was too severe and furthermore requires reconsideration in light of new multi-play rules.