Zone Repop Timer
Changing zone repop settings to (2) so that reboots are no longer required to reset the zone. Replacing this function with a very long timer that should approximate 3 days. (1500) x 2.4 minutes = 3600m / 60m = 60 hours or 3 days. If the high tick counter doesn't cause problems, I feel this solution will abate players need for reboots just to reset a few certain zones. This of course does not affect reboot treasure table, but is still an important step towards a mud not dependent upon reboots.