New weapon procs - molten + possessed
Adding 2 thematic procs for weapons.  

Molten is largely limited to 2h'ers and I attempted to balance it as such.  Bonebreak-style proc but does a wither due to the flame theme.  Flame damage gives a regen to the user, which is stronger than I would have done for a 1h'er to offset the 2h weapon restriction.  It has a couple other bells and whistles like a stronger flame proc if the limb is bleeding (and it consumes the bleed in the process).

Possessed generally has a few damage procs, with a couple rare ones doing a flurry of procs and one having a bloodspike-esque effect.  Not a lot of damage to be done, but there's a chance for procs to bleed the target.  If approved/tweaked, a new item needs to be created to replace the hollow spike, but I'm using the bloodspike item as a placeholder.