Removing "exceptions" to dispel magic for items
There are some legacy exceptions to dispel magic that I don't think make a lot of sense in the current state of the game.  This change will effectively allow all stats to be dispelled from items (previously, stats that affected moves, light, and so on were immune to dispel).  Consequently, this will allow enhance weapon to apply to all weapons.