Preventing weapon dislodge into stunned targets
Currently pierce weapons very frequently get stuck into stunned targets, and this is a bit silly (and annoying).  I think a v2.0 of this change would be to add skill checks to weapon retention/strength/telekinesis/etc. so that you can dislodge the weapon yourself (with bonus damage) from non-stunned targets, but this is a quick and easy quality of life change for now.