Fix Plague of Locusts prayer
Plague of locusts has been broken forever - literally, I remember trying
to use it back in the early 2000s and wondering why it wasn't doing
anything. Feeling personally attacked by this memory, I've fixed it and
it seems to be working as it should now.

Also took it upon myself to refactor it a bit just to clean stuff up and
make the code more descriptive/flow better.

The part that actually fixed the spell was lines 122-126. The swarm was
being summoned successfully but was never actually being made to
attack the victim, and the locusts mob is written to dissipate as soon as
it's no longer involved in a fight - which, before now, was immediately
after being summoned.