Add overload for TBeing::specialAttack
Adds an overload for the TBeing::specialAttack method that accepts a
third param, which acts as a modifier to the attackRound calculation
within. This allows individual physical skills to modify their chance
of success based on the circumstances at the exact moment of each
individual skill attempt.

Converting momentary circumstances into a one-time modifier is done
by determining how many levels worth of bonus or penalty each
situational thing affecting the skill attempt should be worth, multiplying
that number by 16.667 per the balance notes within the TBeing::hits method,
then passing the final value to the specialAttack overload.

That bonus or penalty is then added to the attacker's total offensive
value determined by TBeing::attackRound inside specialAttack.

The rationale for doing it this way is that, currently, the success or
failure of the specialAttack call is heavily dependant on simply the
level difference between attacker and victim. The values are modified
in small ways by various, secondary things (e.g. skills such as cintai
or chivalry) but the vast majority of the final offense value comes
from the attacker's level. Because of this, there's currently very little
players can do to increase their chance of landing many of their
class-defining skills against mobs of any real difficulty.

Many skills currently have some circumstantial modifiers handled
by modifying the value passed to the skill's bSuccess check, which is
undesirable for several reasons. Since this is just a new overload
for specialAttack, skills can be modified to use it on a case-by-case
basis and otherwise nothing will change.