Prevent null dereference on TMonster *&mob
This was a runtime error detected by the undefined behavior sanitizer, and
also caused a seg fault during zone load in my test build compiled with
clang.

Initially I set this code to log when runResetCmdY gets called with TMonster
*&mob == nullptr, but found that it happens a ton of times during zone
loading so the log output was pretty spammy and useless.

After tracing it further it appears the code passes the same TMonster
pointer by ref to all the zone loading functions. When actively loading
a mob it will arrive at the call to runResetCmdY with mob pointing to an
actual TMonster instance. Otherwise it will be a null pointer.

I tested preventing the rs.execute() call at db.cc:3635 completely if
mob was a null pointer and that made it so mobs don't load at all.

So in conclusion, I felt the best way to fix this was just to return from
runResetCmdY immediately if mob == nullptr without logging it.