Prevent bad/undefined value in change pointer
Previous coder forgot to initialize changeWeatherT here, as it is in
other places where a change is passed to the AlterWeather function,
which created some cases where ChangeWeather() would get called
with an undefined/uninitialized value.

This wasn't necessarily causing a bug but was detected as bad/undefined
behavior by UBSan. Initializing it here to CHANGE_NONE fixes the error.