Move skills to be near others in their discipline
Continuation of previous commit. Encamp and garrotte were in the
wrong disciplines, though they stilll showed up correctly in game. This
was causing some bad downcasts to happen which were picked up by
UBSan.

Moving the skills to their correct disciplines' .h files fixed the
errors - this change just moves them to be near the other skills in
those disciplines in skills.cc. The change is superficial, but made
for consistency.