Fix errors and segfault in TMonster::checkSpec
Each mob spec proc has its own function which all share the same
function signature and are stored in an array at the end of
spec_mobs.cc. The vast majority of these functions expect a TObj *
passed in as the final parameter, which is often just a null pointer.

However, a few of the spec proc functions (for instance, shop_keeper)
expect an integer as their final parameter, which represents the
previous room the mob was in.

The existing code used some fairly hacky typecasting to get around
this and make it possible to pass an int where a function was
expecting a TObj*:

- First they static_cast `int was_in` to an intptr_t (which I had to
  change to a long int to get to work in the reinterpret_cast without
  erroring)
- Then they reinterpret cast that resulting value to a TObj*, then
  pass it to the checkSpec function which determines and calls the
  correct spec proc function for <this>.
- HOWEVER: TMonster::checkSpec is inherited from TThing, and the TThing
  version of the function expects a TThing* as the final param. So this
  already reinterpret_cast pseudo-int-pointer was then, inside
  checkSpec, being static_cast BACK to a TObj* before getting passed to
  the mob spec proc function from there
  - In the functions that expect an int value the incoming TObj* is
  static cast back to an int and the original value used

While this appears to have been somehow working all this time, UBSan
was giving runtime errors for bad down/upcasts due to misaligned type
sizes followed by a seg fault, the first static_cast used intptr_t and
when the incoming TThing* in checkSpec was being static_cast back to
a TObj*.

To fix it I added an overload of the TMonster::checkSpec function that
changes the final parameter to an optional TObj* with a default value
of nullptr, since all but one of the calls to TMonster::checkSpec in the
whole codebase already passed a null value for the final param.

This let me remove the static_cast from TThing* to TObj* from inside the
checkSpec function, and between that and changing the first static_cast
to use long int rather than intptr_t (not sure why that worked,
unfortunately) it fixed the runtime errors and seg fault.

However, Cizra mentioned that a better way to do this would be to use
a tagged union type for that final parameter instead, which I'm looking
into to see if it would require updating every single spec proc function
signature. I believe the way I did it here, while not optimal, should at
least work predictably for the time being.