Allow shaman quest mask to act as perma-totem
Credit to Heishiro for the idea.

This commit refactors doSacrifice and task_sacrifice to check for a
worn shaman quest mask before checking for a totem when
using sacrifice.

If a mask is found it takes priority over any totems held. The mask
doesn't lose any durability or suffer any other negative penalty from
the sacrifice.

Added a couple new methods to TBeing to make it easier to check
for/get reference to a held totem or worn shaman mask, since I
needed to make those checks in both doSacrifice and task_sacrifice.

Also removed some unused functions - TThing::sacrificeMe and
TTool::sacrificeMe. The TThing version was never called from anywhere
in the entire codebase, and the TTool version was only called from a
single place, inside doSacrifice. I was able to move all the code out
of the TTool version and straight into doSacrifice, as they were just a
bunch of hard fail checks that didn't depend on the specific item
representing <this> when the function was called.

Tested thoroughly in every way I could think of in my dev instance
and everything works smoothly and the way you'd expect.