Fix memleak in cmd_help.cc
ASan detected a memory leak in cmd_help.cc stemming from the frequent
usage of the mud_str_dup function without proper cleanup.

Instead of trying to troubleshoot the faulty cleanup, this commit
converts the instances of vector<char *> to vector<sstring>, which
modernizes the code and prevents the dynamic memory allocation entirely.