Fix memory leak from garrotte disc mismatch
The garrotte skill was defined in disc_murder.h and is implemented in
disc_murder.cc but was listed as a thief fighting disc skill inside
the main skill array in spell_info.cc.

This mismatch was causing a small memory leak stemming from
skills.cc:889, where garrotte's CDiscipline was being cast as CDMurder
instead of CDThiefFight.

As garrotte makes more sense in the murder disc anyway, changing its
disc in spell_info.cc resolves the memory leak and moves garrotte to the
murder disc in game.