Refactor spells/rituals/prayers command
This is a complete rewrite of the existing
spells/rituals/prayers commands. Prayers was never fully enabled
in the past, though the code was written.

It gets rid of all the rampant char* and sprintf usage in the old
version, some of which were causing memory leaks or potential
memory leaks and tries to use best practices and modern C++
syntax and features whereever possible.

It keeps all the original functionality, while also greatly increasing
the flexibility of the arguments being passed.

It also updates the functions to work well on multiclass characters,
allowing one to view all spells from all classes (or any combination
of classes/spell type thereof) with a single command.