Fix logic to prevent crash when weapon in inv
This proc had some bad logic that allowed it to attempt to trigger
whenever a weapon with the proc was in a character's inventory
when the character used the thief stab skill.

Because the proc was triggering when it shouldn't, it was passing
a garbage value to another call within that was causing a segfault
and crashing the game.

This commit also fixes that the proc was being executed twice
every time stab was used - once from the stab function itself, and
again from the normal check-proc-on-pulse.