Refactor file_to_sstring and add overload
Adds an overload that takes only an sstring filename and returns
a new sstring containing the contents of the file, rather than reading
it into a sstring reference that was passed in. Uses RAII-style
std::ifstream for file access instead of fopen.

Also renames buffer variable to the more descriptive "output".

Refactors the original file_to_sstring function to call this new
overload instead. Greatly simplifies the code and has identical
functionality.