Refactor TBeing::damageLimb to change how limb damage is handled
Limb HP is currently capped at 255 due to it being saved to pfiles.
In the current state of the game, players are able to scrap the limbs
of even high level mobs very easily.
When combined with the updated crit chance code, this means limbs can be
scrapped in a single crit from full HP.
This update changes how limb damage is applied in an attempt to keep the
functionality similar while reducing the ease of scrapping limbs on mobs.