Prevent sacrifice and butcher/skinning from overlapping
If a corpse is being sacrificed at the same time someone else is butchering/skinning it,
and the sacrifice finishes first, the next task pulse for the butcherer/skinner will still
try to reference the now-deleted corpse, causing a heap-use-after-free crash.
These changes prevent those skills from operating simultanesouly on the same corpse.