New Skill: Common
Adds a new skill to advanced adventuring discipline that allows players to mitigate the effects of their accent.  This allows a player to spend their practices to allow other players to better understand them, as opposed to existing skills where the other players need to both spend pracs and hone the various accents to understand others.