Change RUN to CMD in final line
RUN executes the command during the build of the image, which isn't what we want. CMD
executes the command on container startup, which is correct.
For some reason RUN still worked when testing in my local Docker setup, but causes a build failure
when done by the Github Action that builds/pushes a new image on commit.