Update Build and Test GitHub Action and add scons package
This workflow should use optimally use the exact same compilation setup as in the Dockerfile that's used to create the production Docker images. This wasn't the case, so this PR updates it as such.

It also adds the scons package, which was already in the Dockerfile, and is needed here.