Refactor version.h to lib/version.txt, and update usage
The SConstruct file was writing the current commit hash to version.h on every compile. This file was then being
included in two places (main.cc and connect.cc), which was causing those two files to be recompiled on every
build, whether or not they actually changed. This was causing the updates test.yml GitHub workflow to create
a new cache every run, making it less useful.

This commit changes the SConstruct file to instead write the commit hash (and now the commit date as well)
to a text file in the lib folder, version.txt. The hash and date in this file are then read in inside the `main` function
and stored in a global variable, which is then used as needed elsewhere in the code. This prevents the recompilation
issue and allows the test workflow to fully utilize the caching behavior.