Remove references to specific SneezyMUD versions, update credits
As the development of Sneezy is now open source and more of a steady, ongoing process of
adding and updating features, there's likely no more reason to use specific version numbers
when referring to future "eras" of the game. This commit removes all references to SneezyMUD
5.4, instead just using simply SneezyMUD going forward. Versioning will be provided via the
lib/version.txt file, which can be used in the code to display to users the hash and date of the
commit from which the game is currently running.