Completely remove cxxtest library from project
The cxxtest library hasn't had a new release since 2014 and is no longer maintained. It also needs modifications to work with modern versions of SCons. It currently crashes the build in SCons v4.5.2.

Additionally, we're not currently even using the tests in our builds because several of the tests don't work and take forever to complete, as they just timeout repeatedly until finally failing.

Because of these things and to facilitate updating the project to work with recent versions of Ubuntu, SCons, g++, and other libraries I removed cxxtest completely from the project.

Eventually (hopefully sooner rather than later) we'll need to decide on an alternative and get some working tests going again. Google Tests are an industry standard and widely supported, and Boost also has a testing library we could look into.

I left all the tests in the code/code/tests folder for posterity, so they can be used as a reference when we start using a new testing library at some point in the future.