Allowing for the curation of room conditions to allow for an increased regen modifier for Rest and Sleep. This includes a check for these factors in the room, as well as some Thief class-based factors. Also altering Hide to allow for the Rest task, to allow for thieves to regen while remaining vigilant (as opposed to sleeping).
Including an alteration to backstab to allow a bonus for a target who is fighting another opponent to present an edge case where backstab might be more advantageous to use than slit (if conditions to hit are less easily attained). This will begin a series of thief changes that will allow them to exercise skills in a sequence that will make later skill use more likely to land.

Also including a change to some mage spells that will add a wet effect, which ties into an intentional use of cold spells and skills to apply a frostbite effect. This is related to the previous trap-themed commit.