Fix symlink creation to allow "Build and Test" GH Action to run properly
Using CMake to perform the symlink in both release and debug builds is the better option anyway, and uses an absolute path to the GH Action shouldn't have the issue with the verify_boot script not finding the executable. Not sure why I didn't just do it this way in the first place.