cmake: Refactor Dockerfile and CMakeLists to rename binary instead of creating symlink
I fixed a problem with the `Build and Test` workflow by having it symlink the final executable in all builds, but that caused a new issue with the production image build workflow. Having cmake just move the actual binary after the build instead of symlinking should fix both issues at the same time.