Fix build caching in CI workflows (again), upgrade CI dependency versions
Converts the Dockerfile to a single-stage build, where it expects the build to have already been completed and the artifacts to be available in the build context.

Merges the `create-master-cache` workflow into the main `build and publish` workflow that's responsible for creating and pushing new Docker image builds to the container registry.

Doing it like this allows us to cache the build output in a GitHub Actions cache and then reuse that cached build as the base for future runs of the test or publish workflows, which can speed them up considerably when their changes only require recompiling a small number of files (or none, for changes like modifying zonefiles).

However, after some thorough testing it turns out PCH is just completely incompatible with ccache, regardless of what ccache settings are used. But ccache is required to actually utilize the caching behavior during CI, because Ninja only looks at file mtimes to determine what needs to be rebuilt, and when the build cache is restored during the GitHub Actions run, all the cached build objects will have older mtimes than the source files, causing Ninja to just rebuild everything anyway. ccache fixes this by looking at hash values of the source files instead of mtimes, allowing it to correctly determine what actually needs to be rebuilt.

So, to actually get the benefit of ccache during development and the CI actions, PCH just needs to be disabled entirely, which I've set as the default now. Anyone who wants to test PCH locally can still enable it in the build. The biggest downside of not using PCH is that full clean builds will take longer, but incremental builds (more common anyway) should be faster thanks to ccache working now.