Revert user to `sneezy` in Dockerfile to fix volume path issue
In my big Docker overhaul I decided to change the user that runs the game inside the container to `ubuntu` since I was under the impression it was better practice. However, I didn't fully account for the changes everywhere I should have, which led to the volume not being mounted in the right place and no mutable data being available to the game.

After looking into it a bit more I just decided the easiest/cleanest fix was to revert back to using the original `sneezy` user, which fixes the problem.