Fix oversight in Dockerfile causing incorrect volume permissions
When I made the previous change to the Dockerfile to revert the user inside the container
back to `sneezy` I overlooked the fact that the Ubuntu container already includes a built-in
'ubuntu' user with UID 1000. Therefore the `sneezy` user was getting created with UID other than
1000. But the monitor service was configured to set the volume permissions to UID 1000, which
led to the volume being owned by 'ubuntu' user instead of 'sneezy' user, and the game not being able
to read or write to the volume.

Simply renaming the built-in `ubuntu` user to `sneezy` ensures that the `sneezy` user inside the container
has UID 1000, matching the volume ownership set by the monitor service.